rename RoarFooter to Footer
